DIS-37 Phase 2: extract diagnostics into apps/server/src/diagnostics.ts
Lifts the diagnostic-recording subsystem (~280 LOC) out of AgentManager
into a top-level module. Manager now owns no diagnostic state — the
recorder is constructed once in the manager's constructor and the three
public entry points (`maybeCaptureTmuxInventory`, `maybeMaintenanceLogs`,
`captureMissingSessionIncident`) are called by the reconciler directly
rather than going through `this.<helper>`.

What moved:
- 5 timing/size constants (TMUX_INVENTORY_INTERVAL_MS,
  LOG_MAINTENANCE_INTERVAL_MS, MAX_LOG_SIZE_BYTES,
  DIAGNOSTICS_MAX_AGE_MS, SERVER_LOG_MAX_AGE_MS) — were `static readonly`
  on AgentManager, now module-level constants in diagnostics.ts.
- 2 throttle clocks (lastTmuxInventoryAt, lastLogMaintenanceAt) — were
  instance state on AgentManager, now closure variables in the recorder.
  Each call to createDiagnosticsRecorder yields a fresh recorder, which
  is useful for tests and matches how the function is conceptually
  parameterized.
- 9 private methods: diagnosticsRoot, maybeCaptureTmuxInventory,
  captureMissingSessionIncident, maybeMaintenanceLogs, rotateFile,
  copyTruncateFile, deleteOldFiles, detectTmuxServerPid, captureCommand.

Shape choice — factory function rather than class, matching the style
used by the Phase 0 modules (telemetry/persona-reviews/feedback are also
exported functions, not classes). The throttle state is the only
mutable piece, and a closure handles it cleanly. The DiagnosticsRecorder
type captures only the three external entry points; the six internal
helpers (rotateFile etc.) are not part of the public surface.

manager.ts: 2699 → 2416 LOC. Cumulative: 4967 → 2416 (-51%).

Stale fs/promises imports cleaned up: appendFile, open, readdir, rename
no longer needed by manager.ts (only diagnostics used them). os import
also removed.

No new tests this phase: the diagnostic flow is exercised end-to-end
through the reconciler path (covered by existing Playwright tests). The
extracted code is mostly thin wrappers over `fs.rename`/`fs.copyFile`/
`runCommand`, where unit tests would mostly verify Node's behaviour
rather than ours. Throttle behaviour is the one piece worth testing in
isolation; deferring to a follow-up since manager-side tests of
reconciler still cover it.

712/720 server vitest pass (unchanged from Phase 1). 136 e2e pass.
